pyproject.toml
[tool.poetry]
name = "api_pricut"
version = "1.0.0"
description = ""
authors = ["Carlos Aguirre <aguizandre.2803@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
gunicorn = "^22.0.0"
python-decouple = "^3.8"
whitenoise = "^6.6.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
ezdxf = "^1.3.3"
shapely = "^2.0.6"
networkx = "^3.4.2"
django-extensions = "^3.2.3"
deepdiff = "^8.1.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
commitizen = "^3.27.0"
pre-commit = "^3.7.1"
matplotlib = "^3.9.2"

[tool.poetry.group.test.dependencies]
faker = "^25.3.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 83
target-version = ["py312"]
include = "api_pricut/.*\\.pyi?$"

[tool.commitizen]
name = "cz_conventional_commits"
version = "3.27.0"
tag_format = "v$version"

[tool.pytest.ini_options]
pythonpath = "api_pricut"
DJANGO_SETTINGS_MODULE = "settings.environments.testing"
python_files = "test_*.py"
minversion = "8.2.1"
testpaths = [
    "api_pricut/tests",
]
